MassstabY überall verwendet

+++ Orbital.java
@@ -11,10 +11,9 @@ public abstract class Orbital {
     public double x, y, z, v, q;
     public boolean gefunden = false;
     public double Psi;
-    //public static double[][] Fund = new double[100001][4];          //TODO: 100001 prüfen
     private final java.util.Random Zufallsgenerator = new java.util.Random();
-    //TODO: überlegen, wo h wirklich gebraucht wird
-    int h = Rahmen.BildschirmHoehe;
+
+    double h = Rahmen.BildschirmHoehe;
 
     public boolean beobachte(int Nummer) {
 
@@ -25,11 +24,11 @@ public abstract class Orbital {
 
             VersuchsZaehler++;
             // zufälliger Ort
-            x = (Zufallsgenerator.nextDouble() - 0.5) * v;
+            x = (Zufallsgenerator.nextDouble() - 0.5)*v;
             if (Flaeche.Schnitt == 2) x = 0;                          // [-v/2 import java.awt.Color;; v/2] \{0}
-            y = (Zufallsgenerator.nextDouble() - 0.5) * v;
+            y = (Zufallsgenerator.nextDouble() - 0.5)*v;
             if (Flaeche.Schnitt == 3) y = 0;
-            z = (Zufallsgenerator.nextDouble() - 0.5) * v;
+            z = (Zufallsgenerator.nextDouble() - 0.5)*v;
             if (Flaeche.Schnitt == 1) z = 0;
 
             berechneKugelKoordinaten();         //x y z  ->  r phi theta
@@ -59,21 +58,21 @@ public abstract class Orbital {
 
     public void berechneKugelKoordinaten() {
 
-        r = Math.sqrt(x * x + y * y + z * z);                        //Radius
-        if (x == 0) phi = Math.signum(y) * pi / 2;
+        r = Math.sqrt(x*x + y*y + z*z);                        //Radius
+        if (x == 0) phi = Math.signum(y) * pi/2;
         else {
-            phi = Math.atan(y / x);
+            phi = Math.atan(y/x);
             if (x < 0) phi = pi + phi;
         }
         if (r == 0) theta = 0;
-        if (r != 0) theta = Math.acos(z / r);            //theta aus [0;pi]	//phi aus [-pi/2 ; 3pi/2]
+        if (r != 0) theta = Math.acos(z/r);            //theta aus [0;pi]	//phi aus [-pi/2 ; 3pi/2]
     }
 
     public void merkeKoordinaten(int Nummer) {
 
-        Flaeche.Fund[Nummer][1] = x / (v / 2) * ((double) h / 2);    //anpassen an Bildschirmgröße
-        Flaeche.Fund[Nummer][2] = y / (v / 2) * ((double) h / 2);
-        Flaeche.Fund[Nummer][3] = z / (v / 2) * ((double) h / 2);
+        Flaeche.Fund[Nummer][1] = x / (v/2) * (h/2);    //anpassen an Bildschirmgröße
+        Flaeche.Fund[Nummer][2] = y / (v/2) * (h/2);
+        Flaeche.Fund[Nummer][3] = z / (v/2) * (h/2);
         Flaeche.Fund[Nummer][0] = Flaeche.TaktNummer;                     //Zeit der Messung eines Eelektrons
 
         if (Flaeche.Schnitt == 1) {
